% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_main.R
\name{catecv}
\alias{catecv}
\title{Cross-validation of the conditional average treatment effect (CATE) score for count, survival or continuous outcomes}
\usage{
catecv(
  response,
  data,
  score.method,
  cate.model,
  ps.model,
  ps.method = "glm",
  init.model = NULL,
  initial.predictor.method = NULL,
  ipcw.model = NULL,
  ipcw.method = "breslow",
  minPS = 0.01,
  maxPS = 0.99,
  followup.time = NULL,
  tau0 = NULL,
  higher.y = TRUE,
  prop.cutoff = seq(0.5, 1, length = 6),
  prop.multi = c(0, 1/3, 2/3, 1),
  abc = TRUE,
  train.prop = 3/4,
  cv.n = 10,
  error.max = 0.1,
  max.iter = 5000,
  surv.min = 0.025,
  xvar.smooth.score = NULL,
  xvar.smooth.init = NULL,
  tree.depth = 2,
  n.trees.rf = 1000,
  n.trees.boosting = 200,
  B = 3,
  Kfold = 5,
  error.maxNR = 0.001,
  max.iterNR = 150,
  tune = c(0.5, 2),
  seed = NULL,
  plot.gbmperf = TRUE,
  verbose = 0
)
}
\arguments{
\item{response}{A string describing the type of outcome in the data.
Allowed values include "count" (see \code{\link{catecvcount}()}), "survival"
(see \code{\link{catecvsurv}()}) and "continuous" (see
\code{\link{catecvmean}()}) .}

\item{data}{A data frame containing the variables in the outcome, propensity score, and inverse
probability of censoring models (if specified); a data frame with \code{n} rows (1 row per observation).}

\item{score.method}{A vector of one or multiple methods to estimate the CATE score.
Allowed values are: \code{'boosting'}, \code{'twoReg'}, \code{'contrastReg'}, \code{'poisson'} (count and survival outcomes only),
\code{'randomForest'} (survival, continuous outcomes only), \code{negBin} (count outcomes only), \code{'gam'} (continuous outcomes only),
\code{'gaussian'} (continuous outcomes only).}

\item{cate.model}{A formula describing the outcome model to be fitted.
The outcome must appear on the left-hand side. For survival outcomes, a
\code{Surv} object must be used to describe the outcome.}

\item{ps.model}{A formula describing the propensity score (PS) model to be fitted. The treatment must
appear on the left-hand side. The treatment must be a numeric vector coded as 0/1.
If data are from a randomized controlled trial, specify \code{ps.model = ~1} as an intercept-only model.}

\item{ps.method}{A character value for the method to estimate the propensity score.
Allowed values include one of:
\code{'glm'} for logistic regression with main effects only (default), or
\code{'lasso'} for a logistic regression with main effects and LASSO penalization on
two-way interactions (added to the model if interactions are not specified in \code{ps.model}).
Relevant only when \code{ps.model} has more than one variable.}

\item{init.model}{A formula describing the initial predictor model. The outcome must appear on the left-hand side.
It must be specified when \code{score.method = contrastReg} or \code{twoReg}.}

\item{initial.predictor.method}{A character vector for the method used to get initial
outcome predictions conditional on the covariates specified in \code{cate.model}. Only applies
when \code{score.method} includes \code{'twoReg'} or \code{'contrastReg'}. Allowed values include one of
\code{'randomForest'} (survival outcomes only), \code{'boosting'}, \code{'logistic'}
(survival outcomes only, fast), \code{'poisson'} (count outcomes only, fast), \code{'gaussian'} (continuous outcomes only) and
\code{'gam'} (count and continuous outcomes only). Default is \code{NULL}, which assigns \code{'boosting'}
for count outcomes and \code{'randomForest'} for survival outcomes.}

\item{ipcw.model}{A formula describing the inverse probability of censoring weighting (IPCW)
model to be fitted. The left-hand side must be empty. Only applies for survival outcomes.
Default is \code{NULL}, which corresponds to specifying the IPCW with the same covariates
as the outcome model \code{cate.model}, plus the treatment.}

\item{ipcw.method}{A character value for the censoring model. Only applies for survival
outcomes. Allowed values are: \code{'breslow'} (Cox regression with Breslow estimator of t
he baseline survivor function), \code{'aft (exponential)'}, \code{'aft (weibull)'},
\code{'aft (lognormal)'} or \code{'aft (loglogistic)'} (accelerated failure time model
with different distributions for y variable). Default is \code{'breslow'}.}

\item{minPS}{A numerical value (in [0, 1]) below which estimated propensity scores should be
truncated. Default is \code{0.01}.}

\item{maxPS}{A numerical value (in (0, 1]) above which estimated propensity scores should be
truncated. Must be strictly greater than \code{minPS}. Default is \code{0.99}.}

\item{followup.time}{A column name in \code{data} specifying the maximum follow-up time,
interpreted as the potential censoring time. Only applies for survival outcomes.
Default is \code{NULL}, which corresponds to unknown potential censoring time.}

\item{tau0}{The truncation time for defining restricted mean time lost. Only applies for
survival outcomes. Default is \code{NULL}, which corresponds to setting the truncation time as the
maximum survival time in the data.}

\item{higher.y}{A logical value indicating whether higher (\code{TRUE}) or
lower (\code{FALSE}) values of the outcome are more desirable. Default is \code{TRUE}.}

\item{prop.cutoff}{A vector of numerical values (in (0, 1]) specifying percentiles of the
estimated log CATE scores to define nested subgroups. Each element represents the cutoff to
separate observations in nested subgroups (below vs above cutoff).
The length of \code{prop.cutoff} is the number of nested subgroups.
An equally-spaced sequence of proportions ending with 1 is recommended.
Default is \code{seq(0.5, 1, length = 6)}.}

\item{prop.multi}{A vector of numerical values (in [0, 1]) specifying percentiles of the
estimated log CATE scores to define mutually exclusive subgroups.
It should start with 0, end with 1, and be of \code{length(prop.multi) > 2}.
Each element represents the cutoff to separate the observations into
\code{length(prop.multi) - 1} mutually exclusive subgroups.
Default is \code{c(0, 1/3, 2/3, 1)}.}

\item{abc}{A logical value indicating whether the area between curves (ABC) should be calculated
at each cross-validation iterations, for each \code{score.method}. Default is \code{TRUE}.}

\item{train.prop}{A numerical value (in (0, 1)) indicating the proportion of total data used
for training. Default is \code{3/4}.}

\item{cv.n}{A positive integer value indicating the number of cross-validation iterations.
Default is \code{10}.}

\item{error.max}{A numerical value > 0 indicating the tolerance (maximum value of error)
for the largest standardized absolute difference in the covariate distributions or in the
doubly robust estimated rate ratios between the training and validation sets. This is used
to define a balanced training-validation splitting. Default is \code{0.1}.}

\item{max.iter}{A positive integer value indicating the maximum number of iterations when
searching for a balanced training-validation split. Default is \code{5,000}.}

\item{surv.min}{Lower truncation limit for the probability of being censored.
It must be a positive value and should be chosen close to 0. Only applies for survival
outcomes. Default is \code{0.025}.}

\item{xvar.smooth.score}{A vector of characters indicating the name of the variables used as
the smooth terms if \code{score.method = 'gam'}. The variables must be selected
from the variables listed in \code{cate.model}.}

\item{xvar.smooth.init}{A vector of characters indicating the name of the variables used as
the smooth terms if \code{initial.predictor.method = 'gam'}. The variables must be selected
from the variables listed in \code{init.model}.
Default is \code{NULL}, which uses all variables in \code{init.model}.}

\item{tree.depth}{A positive integer specifying the depth of individual trees in boosting
(usually 2-3). Used only if \code{score.method = 'boosting'} or
if \code{initial.predictor.method = 'boosting'} with \code{score.method = 'twoReg'} or
\code{'contrastReg'}. Default is 2.}

\item{n.trees.rf}{A positive integer specifying the maximum number of trees in random forest.
Used if \code{score.method = 'ranfomForest'} or if \code{initial.predictor.method = 'randomForest'}
with \code{score.method = 'twoReg'} or \code{'contrastReg'}. Only applies for survival outcomes.
Default is \code{1000}.}

\item{n.trees.boosting}{A positive integer specifying the maximum number of trees in boosting
(usually 100-1000). Used if \code{score.method = 'boosting'} or
if \code{initial.predictor.method = 'boosting'} with \code{score.method = 'twoReg'} or
\code{'contrastReg'}. Default is \code{200}.}

\item{B}{A positive integer specifying the number of time cross-fitting is repeated in
\code{score.method = 'twoReg'} and \code{'contrastReg'}. Default is \code{3}.}

\item{Kfold}{A positive integer specifying the number of folds used in cross-fitting
to partition the data in \code{score.method = 'twoReg'} and \code{'contrastReg'}.
Default is \code{5}.}

\item{error.maxNR}{A numerical value > 0 indicating the minimum value of the mean absolute
error in Newton Raphson algorithm. Used only if \code{score.method = 'contrastReg'}.
Default is \code{0.001}.}

\item{max.iterNR}{A positive integer indicating the maximum number of iterations in the
Newton Raphson algorithm. Used only if \code{score.method = 'contrastReg'}.
Default is \code{150}.}

\item{tune}{A vector of 2 numerical values > 0 specifying tuning parameters for the
Newton Raphson algorithm. \code{tune[1]} is the step size, \code{tune[2]} specifies a
quantity to be added to diagonal of the slope matrix to prevent singularity.
Used only if \code{score.method = 'contrastReg'}. Default is \code{c(0.5, 2)}.}

\item{seed}{An optional integer specifying an initial randomization seed for reproducibility.
Default is \code{NULL}, corresponding to no seed.}

\item{plot.gbmperf}{A logical value indicating whether to plot the performance measures in
boosting. Used only if \code{score.method = 'boosting'} or if \code{score.method = 'twoReg'}
or \code{'contrastReg'} and \code{initial.predictor.method = 'boosting'}. Default is \code{TRUE}.}

\item{verbose}{An integer value indicating what kind of intermediate progress messages should
be printed. \code{0} means no outputs. \code{1} means only progress bar and run time.
\code{2} means progress bar, run time, and all errors and warnings. Default is \code{0}.}
}
\value{
For count response, see description of outputs in \code{\link{catecvcount}()}.
For survival response, see description of outputs in \code{\link{catecvsurv}()}.
For continuous response, see description of outputs in \code{\link{catecvmean}()}.
}
\description{
Provides (doubly robust) estimation of the average treatment effect (ATE) for count, survival or continuous
outcomes in nested and mutually exclusive subgroups of patients defined by an estimated conditional
average treatment effect (CATE) score via cross-validation (CV).
}
\details{
For count response, see details in \code{\link{catecvcount}()}.
For survival response, see details in \code{\link{catecvsurv}()}.
For continuous response, see details in \code{\link{catecvmean}()}.
}
\examples{
\dontrun{
output_catecv <- catecv(response = "count",
                        data = countExample,
                        score.method = "poisson",
                        cate.model = y ~ age + female + previous_treatment + previous_cost +
                                     previous_number_relapses + offset(log(years)),
                        ps.model = trt ~ age + previous_treatment,
                        higher.y = FALSE,
                        cv.n = 5,
                        plot.gbmperf = FALSE,
                        seed = 999)

# Try:
plot(x = output_catecv, ylab = "RMTL ratio of drug1 vs drug0 in each subgroup")
boxplot(x = output_catecv, ylab = "RMTL ratio of drug1 vs drug0 in each subgroup")
abc(x = output_catecv)

# Survival outcome
tau0 <- with(survivalExample,
             min(quantile(y[trt == "drug1"], 0.95), quantile(y[trt == "drug0"], 0.95)))

output_catecv2 <- catecv(response = "survival",
                         data = survivalExample,
                         score.method = c("poisson", "randomForest"),
                         cate.model = survival::Surv(y, d) ~ age +
                                                     female +
                                                     previous_cost +
                                                     previous_number_relapses,
                         ps.model = trt ~ age + previous_treatment,
                         initial.predictor.method = "randomForest",
                         ipcw.model = ~ age + previous_cost + previous_treatment,
                         followup.time = NULL,
                         tau0 = tau0,
                         higher.y = TRUE,
                         surv.min = 0.025,
                         cv.n = 5,
                         seed = 999,
                         plot.gbmperf = FALSE)

# Try:
plot(x = output_catecv2, ylab = "RMTL ratio of drug1 vs drug0 in each subgroup")
boxplot(x = output_catecv2, ylab = "RMTL ratio of drug1 vs drug0 in each subgroup")
abc(x = output_catecv2)


# Continuous outcome
output_catecv3 <- catecv(response = "continuous",
                         data = meanExample,
                         score.method = c("gaussian", "twoReg"),
                         cate.model = y ~ age +
                                 previous_treatment +
                                 previous_cost +
                                 previous_status_measure,
                         ps.model = trt ~ previous_status_measure,
                         init.model = y ~ age +
                                 previous_treatment +
                                 previous_cost +
                                 previous_status_measure,
                         initial.predictor.method = "gaussian",
                         higher.y = FALSE,
                         cv.n = 5,
                         xvar.smooth.score = c("age", "previous_cost"),
                         seed = 999,
                         plot.gbmperf = FALSE)

# Try:
plot(x = output_catecv3)
boxplot(x = output_catecv3)
abc(x = output_catecv3)
}

}
\references{
Yadlowsky, S., Pellegrini, F., Lionetto, F., Braune, S., & Tian, L. (2020).
\emph{Estimation and validation of ratio-based conditional average treatment effects using
observational data. Journal of the American Statistical Association, 1-18.}
\url{https://www.tandfonline.com/doi/full/10.1080/01621459.2020.1772080}
}
\seealso{
\code{\link{catefit}()} function and \code{\link{boxplot}()}, \code{\link{abc}} methods for
\code{"precmed"} objects.
}
