% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_continuous.R
\name{meanCatch}
\alias{meanCatch}
\title{Catch errors and warnings when estimating the ATEs in the nested subgroup for continuous data}
\usage{
meanCatch(fun)
}
\arguments{
\item{fun}{The drsurv function...}
}
\value{
A list containing
}
\description{
Storing the errors and warnings that occurred when estimating the ATEs in the nested subgroups.
If there are no errors and no warnings, the estimated mean difference is provided.
If there are warnings but no errors, the estimated mean difference is provided with a warning attribute set.
If there are errors, the NA values are returned for mean difference. A error attribute set is also provided.
}
