% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATE_generic.R
\name{plot.atefit}
\alias{plot.atefit}
\title{Histogram of bootstrap estimates}
\usage{
\method{plot}{atefit}(x, bins = 50, alpha = 0.7, ...)
}
\arguments{
\item{x}{An object of class \code{"atefit"}.}

\item{bins}{Number of bins.}

\item{alpha}{Opacity}

\item{...}{Other parameters

Create a histogram displaying the distribution of the bootstrap estimates.
The red vertical reference line represents the final estimate.}
}
\description{
Histogram of bootstrap estimates
}
