% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATE_generic.R
\name{print.atefit}
\alias{print.atefit}
\title{Print function for atefit}
\usage{
\method{print}{atefit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"atefit"}.}

\item{...}{Other parameters}
}
\description{
Print function for atefit
}
