% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATE.R
\name{print.catefit}
\alias{print.catefit}
\title{Print function for atefit}
\usage{
\method{print}{catefit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"catefit"}.}

\item{...}{Other parameters}
}
\value{
No return value
}
\description{
Print function for atefit
}
\details{
Display the estimated treatment effects for survival outcomes (log
restricted mean time lost ratio and log hazard ratio) and count outcomes
(the log rate ratio).
}
\author{
Thomas Debray
}
