% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survivalExample}
\alias{survivalExample}
\title{Simulated data with survival outcome}
\format{
A dataframe with 4000 rows (patients) and 9 variables:
\describe{
  \item{age}{age at baseline, centered to 48 years old, in years}
  \item{female}{sex, 0 for male, 1 for female}
  \item{previous_treatment}{previous treatment, "drugA", "drugB", or "drugC"}
  \item{previous_cost}{previous medical cost, in US dollars}
  \item{previous_number_symptoms}{previous number of symptoms, "0", "1", or ">=2"}
  \item{previous_number_relapses}{previous number of relapses}
  \item{trt}{current treatment, "drug0" or "drug1"}
  \item{y}{time to first relapse or censoring}
  \item{d}{event indicator, 1: relapse, 0: censored}
}
}
\usage{
data(survivalExample)
}
\description{
A dataset containing a time-to-event outcome, an event indicator, treatment group, and 6 baseline covariates
}
\examples{
data(survivalExample)
str(survivalExample)
}
\keyword{datasets}
