% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATE.R
\name{plot.atefit}
\alias{plot.atefit}
\title{Histogram of bootstrap estimates}
\usage{
\method{plot}{atefit}(x, bins, alpha = 0.7, title = waiver(), theme = theme_classic(), ...)
}
\arguments{
\item{x}{An object of class \code{"atefit"}.}

\item{bins}{Number of bins}

\item{alpha}{Opacity}

\item{title}{The text for the title}

\item{theme}{Defaults to \code{theme_classic()}. Other options include \code{theme_grey()}, \code{theme_bw()}, \code{theme_light()}, \code{theme_dark()}, and \code{theme_void()}}

\item{...}{Other parameters}
}
\value{
A plot of the class \code{ggplot}, displaying the estimated ATE across
the bootstrap samples
}
\description{
Histogram of bootstrap estimates
}
\details{
Create a histogram displaying the distribution of the bootstrap estimates.
The red vertical reference line represents the final estimate.
}
\author{
Thomas Debray
}
